Break out rating to non static, and get it wrapped, etc... properly for jaxb
