+++ src/main/java/Rating.java
@@ -1,0 +1,2 @@
+public class Rating {
+}
