+++ src/main/java/org/cyclonedx/model/vulnerability/Rating.java
@@ -1,0 +1,73 @@
+/*
+ * This file is part of CycloneDX Core (Java).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * Copyright (c) Steve Springett. All Rights Reserved.
+ */
+package org.cyclonedx.model.vulnerability;
+
+import com.thoughtworks.xstream.annotations.XStreamAlias;
+import jakarta.xml.bind.annotation.XmlElement;
+import org.cyclonedx.model.vulnerability.Vulnerability1_0.Score;
+import org.cyclonedx.model.vulnerability.Vulnerability1_0.ScoreSource;
+import org.cyclonedx.model.vulnerability.Vulnerability1_0.Severity;
+
+@XStreamAlias("rating")
+public class Rating
+{
+  private Score score;
+
+  private Severity severity;
+
+  private ScoreSource method;
+
+  private String vector;
+
+  @XmlElement(name = "score", namespace = Vulnerability1_0.NAMESPACE_URI)
+  public Score getScore() {
+    return score;
+  }
+
+  public void setScore(final Score score) {
+    this.score = score;
+  }
+
+  @XmlElement(name = "severity", namespace = Vulnerability1_0.NAMESPACE_URI)
+  public Severity getSeverity() {
+    return severity;
+  }
+
+  public void setSeverity(final Severity severity) {
+    this.severity = severity;
+  }
+
+  @XmlElement(name = "method", namespace = Vulnerability1_0.NAMESPACE_URI)
+  public ScoreSource getMethod() {
+    return method;
+  }
+
+  public void setMethod(final ScoreSource method) {
+    this.method = method;
+  }
+
+  @XmlElement(name = "vector", namespace = Vulnerability1_0.NAMESPACE_URI)
+  public String getVector() {
+    return vector;
+  }
+
+  public void setVector(final String vector) {
+    this.vector = vector;
+  }
+}
