+++ src/main/java/org/cyclonedx/model/vulnerability/Vulnerability1_0.java
@@ -25,6 +25,7 @@ import java.io.Serializable;
 import com.google.gson.annotations.SerializedName;
 import jakarta.xml.bind.annotation.XmlAttribute;
 import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlElementWrapper;
 import jakarta.xml.bind.annotation.XmlRootElement;
 
 import com.thoughtworks.xstream.annotations.XStreamAlias;
@@ -145,7 +146,8 @@ public class Vulnerability1_0
       this.cwes = cwes;
     }
 
-    @XmlElement(name = "ratings", namespace = NAMESPACE_URI)
+    @XmlElementWrapper(name = "ratings")
+    @XmlElement(name = "rating", namespace = NAMESPACE_URI)
     public List<Rating> getRatings() {
       return ratings;
     }
@@ -206,51 +208,7 @@ public class Vulnerability1_0
       }
     }
 
-    @XStreamAlias("rating")
-    public static class Rating {
-      private Score score;
-      private Severity severity;
-      private ScoreSource method;
-      private String vector;
-
-      @XmlElement(name = "score", namespace = NAMESPACE_URI)
-      public Score getScore() {
-        return score;
-      }
-
-      public void setScore(final Score score) {
-        this.score = score;
-      }
-
-      @XmlElement(name = "severity", namespace = NAMESPACE_URI)
-      public Severity getSeverity() {
-        return severity;
-      }
-
-      public void setSeverity(final Severity severity) {
-        this.severity = severity;
-      }
-
-      @XmlElement(name = "method", namespace = NAMESPACE_URI)
-      public ScoreSource getMethod() {
-        return method;
-      }
-
-      public void setMethod(final ScoreSource method) {
-        this.method = method;
-      }
-
-      @XmlElement(name = "vector", namespace = NAMESPACE_URI)
-      public String getVector() {
-        return vector;
-      }
-
-      public void setVector(final String vector) {
-        this.vector = vector;
-      }
-    }
-
-    @XStreamAlias("score")
+  @XStreamAlias("score")
     public static class Score {
       private Double base;
       private Double impact;
